Map leader to space and add hls! leader mapping
I prefer having search highlighting as a mode rather than having it on
and clearing it all the time with :nohlsearch. Usually I don't need it
as it makes search more ergonomic to use for navigation. For cases when
I want to search for something with highlighting (e.g. building a search
pattern) it makes more sense to toggle the mode.