Append history when shell exits
Do not interleave history of different sessions to have consistent view
of what has happened*. Do not remove duplicated commands for the same
reason.

*You can define consistent history in two different ways: one will give
you consistent history of actions performed on the machine the other -
consistent view of history of actions performed in a session by
serializing different sessions in the history.