Do not use ~ in $PATH
While bash expands ~ in $PATH correctly other shells does not. And $PATH
exported by ~/.profile is used by other shells, not just bash. For
example when a program uses exec() to run a command it will fail to find
any binaries in parts of the $PATH that use ~ (watchexec/watchexec#121).