Fix makefile for directories
When link name passed to ln is a directory if will be used as the
destination for the link file. This will cause another symlink to be
created inside of the directory that the symlink targets on the
subsequent run of make, resulting in the directory containing a symlink
to itself.